implement special case for "stare mode" (repeatedly retune same frequency).

+++ include/gnuradio/iqtlabs/retune_fft.h
@@ -228,12 +228,12 @@ public:
    * class. iqtlabs::retune_fft::make is the public interface for
    * creating new instances.
    */
-  static sptr make(const std::string &tag, int vlen, int nfft,
+  static sptr make(const std::string &tag, size_t vlen, size_t nfft,
                    uint64_t samp_rate, uint64_t freq_start, uint64_t freq_end,
-                   int tune_step_hz, int tune_step_fft, int skip_tune_step_fft,
-                   double fft_min, double fft_max, const std::string &sdir,
-                   uint64_t write_step_fft, double bucket_range,
-                   const std::string &tuning_ranges,
+                   uint64_t tune_step_hz, uint64_t tune_step_fft,
+                   uint64_t skip_tune_step_fft, double fft_min, double fft_max,
+                   const std::string &sdir, uint64_t write_step_fft,
+                   double bucket_range, const std::string &tuning_ranges,
                    const std::string &description, uint64_t rotate_secs);
 };
